Fixing double issue

+++ starters/math/Juicedratics.java
@@ -4,14 +4,14 @@ import java.util.Scanner;
 public class Juicedratics {
 	public static void main(String[] args) throws IOException {
 		Scanner scanner = new Scanner(System.in);
-		int a = scanner.nextInt();
-		int b = scanner.nextInt();
-		int c = scanner.nextInt();
+		double a = scanner.nextDouble();
+		double b = scanner.nextDouble();
+		double c = scanner.nextDouble();
 		
 		findSolutions(a, b, c);
 	}
 	
-	public static void findSolutions(int a, int b, int c) {
+	public static void findSolutions(double a, double b, double c) {
 		/*
 		 * Solve the quadratic equation here.
 		 * Set the variables to their appropriate values,
